requirements.txt
flask
requests
python-telegram-bot
langchain
langgraph
langchain-openai
sqlite-utils